+++ src/main/java/ru/kata/spring/boot_security/demo/dao/UserDao.java
@@ -10,9 +10,9 @@ public interface UserDao {
 
     void updateUser(User user);
 
-    void removeUser(int id);
+    void removeUser(Long id);
 
-    User getUserById(int id);
+    User getUserById(Long id);
 
     List<User> getListUsers();
